config: move redirector logic to module
