+++ site/zenodo_rdm/config.py
@@ -6,7 +6,20 @@
 # under the terms of the MIT License; see LICENSE file for more details.
 """Custom code config."""
 
-import os
+from .redirector import (
+    communities_detail_view_function,
+    communities_records_search,
+    communities_requests_view_function,
+    communities_settings_view_function,
+    deposit_view_function,
+    record_export_view,
+    record_file_download_view,
+    record_view_function,
+    redirect_deposit_new_view,
+    redirect_deposit_own_view,
+    redirect_record_file_preview_view,
+    search_view_function,
+)
 
 # I18N_TRANSLATIONS_PATHS = [os.path.abspath("./site/zenodo_rdm/translations")]
 
@@ -163,228 +176,20 @@ SUPPORT_EMAIL_CONFIRM_BODY_TEMPLATE = "zenodo_rdm/email_confirm_body.html"
 #: Email title template.
 SUPPORT_EMAIL_CONFIRM_TITLE_TEMPLATE = "zenodo_rdm/email_confirm_title.html"
 
+# Search query of recent uploads
+# Defaults to newest records search
+ZENODO_FRONTPAGE_RECENT_UPLOADS_QUERY = "/api/records?sort=newest&size=10"
+
 # Citations
+# =========
 ZENODO_RECORDS_UI_CITATIONS_ENDPOINT = (
     "https://zenodo-broker-qa.web.cern.ch/api/relationships"
 )
 
 ZENODO_RECORDS_UI_CITATIONS_ENABLE = False
 
-
-ZENODO_TYPE_SUBTYPE_LEGACY = {
-    "publications": "publication",
-    "books": "publication.book",
-    "books-sections": "publication.section",
-    "conference-papers": "publication.conferencepaper",
-    "journal-articles": "publication.article",
-    "patents": "publication.patent",
-    "preprints": "publication.preprint",
-    "deliverable": "publication.deliverable",
-    "milestone": "publication.milestone",
-    "proposal": "publication.proposal",
-    "reports": "publication.report",
-    "theses": "publication.thesis",
-    "technical-notes": "publication.technicalnote",
-    "working-papers": "publication.workingpaper",
-    "other-publications": "publication.other",
-    "posters": "poster",
-    "presentations": "presentation",
-    "datasets": "dataset",
-    "images": "image",
-    "figures": "image.figure",
-    "drawings": "image.drawing",
-    "diagrams": "image.diagram",
-    "photos": "image.photo",
-    "other-images": "image.other",
-    "videos": "video",
-    "software": "software",
-    "lessons": "lesson",
-    "physicalobject": "physicalobject",
-    "workflows": "workflow",
-    "other": "other",
-}
-
-
-from flask import request, url_for
-
-
-def communities_detail_view_function():
-    """Implements redirector view function for communities detail.
-
-    The following routes are redirected as follows:
-        - /communities/about/<id>/ -> GET /communities/<pid_value>
-        - /collection/user-<id> -> GET /communities/<pid_value>
-        - /communities/<community_id>/about -> GET /communities/<pid_value>
-
-    :return: url for the view 'invenio_app_rdm_communities.communities_detail'
-    :rtype: str
-    """
-    _id = request.view_args.get("id", request.view_args.get("community_id"))
-    values = {"pid_value": _id}
-    target = url_for("invenio_app_rdm_communities.communities_detail", **values)
-    return target
-
-
-def communities_settings_view_function():
-    """Implements redirector view function for communities settings.
-
-    The following routes are redirected as follows:
-        - /communities/<community_id>/edit -> GET /communities/<community_id>/settings
-
-    :return: url for the view 'invenio_communities.communities_settings'
-    :rtype: str
-    """
-    _id = request.view_args.get("id", request.view_args.get("community_id"))
-    values = {"pid_value": _id}
-    target = url_for("invenio_communities.communities_settings", **values)
-    return target
-
-
-def communities_requests_view_function():
-    """Implements redirector view function for communities requests.
-
-    The following routes are redirected as follows:
-        - /communities/<community_id>/curate -> GET /communities/<community_id>/settings
-
-    :return: url for the view 'invenio_communities.communities_requests'
-    :rtype: str
-    """
-    _id = request.view_args.get("id", request.view_args.get("community_id"))
-    values = {"pid_value": _id}
-    target = url_for("invenio_communities.communities_requests", **values)
-    return target
-
-
-def communities_records_search():
-    """Implements redirector view function for communities records search.
-
-    The following routes are redirected as follows:
-        - /communities/<string:community_id>/search -> GET /communities/<pid_value>/q=<query>
-
-    :return: url for the view 'invenio_communities.communities_detail'
-    :rtype: str
-    """
-    _id = request.view_args.get("id", request.view_args.get("community_id"))
-    values = {"pid_value": _id}
-    _q = request.args.get("q", "")
-    url = url_for("invenio_app_rdm_communities.communities_detail", **values)
-    target = f"{url}?q={_q}"
-    return target
-
-
-def search_view_function():
-    """Implements redirector view function for search ui.
-
-    The following routes are redirected as follows:
-        - /collection/<type> -> GET /search?q=<type>
-
-    :return: url for the view 'invenio_search_ui.search'
-    :rtype: str
-    """
-    _type = request.view_args["type"]
-    legacy_value = ZENODO_TYPE_SUBTYPE_LEGACY.get(_type)
-    values = {"q": f"metadata.resource_type.id:{legacy_value}"}
-    target = url_for("invenio_search_ui.search", **values)
-    return target
-
-
-def deposit_view_function():
-    """Implements redirector view function for deposit edit.
-
-    The following routes are redirected as follows:
-        - /deposit/<pid_value>  -> GET /uploads/<pid_value>
-
-    :return: url for the view 'invenio_app_rdm_records.deposit_edit'
-    :rtype: str
-    """
-    values = request.view_args
-    target = url_for("invenio_app_rdm_records.deposit_edit", **values)
-    return target
-
-
-def record_view_function():
-    """Implements redirector view function for record detail.
-
-    The following routes are redirected as follows:
-        - /record/<pid_value>', -> GET /records/<pid_value>
-
-    :return: url for the view 'invenio_app_rdm_records.record_detail'
-    :rtype: str
-    """
-    values = request.view_args
-    target = url_for("invenio_app_rdm_records.record_detail", **values)
-    return target
-
-
-def record_export_view():
-    """Implements redirector view function for record export.
-
-    The following routes are redirected as follows:
-        -  /record/<pid_value>/export/<export_format> -> GET /records/<pid_value>/export/<export_format>
-
-    :return: url for the view 'invenio_app_rdm_records.record_export'
-    :rtype: str
-    """
-    values = request.view_args
-    target = url_for("invenio_app_rdm_records.record_export", **values)
-    return target
-
-
-def record_file_download_view():
-    """Implements redirector view function for record export.
-
-    The following routes are redirected as follows:
-        -  /record/<pid_value>/files/<filename> -> GET /records/<pid_value>/files/<filename>
-
-    :return: url for the view 'invenio_app_rdm_records.record_file_download'
-    :rtype: str
-    """
-    values = request.view_args
-    target = url_for("invenio_app_rdm_records.record_file_download", **values)
-    return target
-
-
-def redirect_record_file_preview_view():
-    """Implements redirector view function for record file preview.
-
-    The following routes are redirected as follows:
-        -  /record/<pid_value>/preview/<filename> -> /records/<pid_value>/preview/<filename>
-
-    :return: url for the view 'invenio_app_rdm_records.record_file_preview'
-    :rtype: str
-    """
-    values = request.view_args
-    target = url_for("invenio_app_rdm_records.record_file_preview", **values)
-    return target
-
-
-def redirect_deposit_own_view():
-    """Implements redirector view function for user uploads.
-
-    The following routes are redirected as follows:
-        -  /deposit -> GET /me/uploads
-
-    :return: url for the view 'invenio_app_rdm_users.uploads'
-    :rtype: str
-    """
-    _q = "&is_published=false"
-    url = url_for("invenio_app_rdm_users.uploads")
-    target = f"{url}?q={_q}"
-    return target
-
-
-def redirect_deposit_new_view():
-    """Implements redirector view function for deposit creation.
-
-    The following routes are redirected as follows:
-        -  /deposit/new -> GET /uploads/new
-
-    :return: url for the view 'invenio_app_rdm_records.deposit_create'
-    :rtype: str
-    """
-    target = url_for("invenio_app_rdm_records.deposit_create")
-    return target
-
+# Redirection
+# ===========
 
 REDIRECTOR_RULES = {
     "redirect_communities_about_legacy": {
